+++ pkg/fn/runtime/runner_test.go
@@ -14,7 +14,7 @@
 
 // Package pipeline provides struct definitions for Pipeline and utility
 // methods to read and write a pipeline resource.
-package fnruntime
+package runtime
 
 import (
 	"bytes"
