+++ pkg/fn/runtime/wasmtime_unsupported.go
@@ -14,7 +14,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package fnruntime
+package runtime
 
 // Stub functions for running without wasmtime support compiled in.
 // wasmtime requires cgo, which is not always a viable option.
